BACKEND/requirements.txt
# ── Web framework ────────────────────────────────────────────────────────────
fastapi
uvicorn[standard]

# ── Base de datos ────────────────────────────────────────────────────────────
sqlalchemy
psycopg2-binary
pgvector
python-dotenv

# ── Validación ───────────────────────────────────────────────────────────────
pydantic[email]
email-validator

# ── Autenticación ────────────────────────────────────────────────────────────
python-jose[cryptography]
cryptography
passlib[bcrypt]
bcrypt
python-multipart

# ── Reconocimiento facial ─────────────────────────────────────────────────────
insightface
onnxruntime
opencv-python-headless
numpy

# ── Notificaciones ─────────────────────────────────────────────────────────────
twilio